Huge clean-up, no new code. Just organizing all files and placing Composables into their own files.
