+++ app/src/main/java/com/heyzeusv/solitaire/ui/Card.kt
@@ -1,4 +1,4 @@
-package com.heyzeusv.solitaire
+package com.heyzeusv.solitaire.ui
 
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.Image
@@ -22,20 +22,12 @@ import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.heyzeusv.solitaire.R
+import com.heyzeusv.solitaire.data.Card
+import com.heyzeusv.solitaire.data.cardsMap
 import com.heyzeusv.solitaire.ui.theme.SolitaireTheme
+import com.heyzeusv.solitaire.util.Suits
 import com.heyzeusv.solitaire.util.autosizetext.AutoSizeText
-
-// card value to their display value
-val cardsMap = mapOf(0 to "A", 1 to "2", 2 to "3" , 3 to "4", 4 to "5", 5 to "6", 6 to "7", 7 to "8", 8 to "9", 9 to "10", 10 to "J", 11 to "Q", 12 to "K")
-/**
- *  Data class containing information for individual cards. [value] ranges from 0 to 13 which
- *  corresponds to the values 1 to 10, A, J, Q, K. [suit] is one of 4 possible values from the
- *  [Suits] enum class. [faceUp] determines if the user can see the card.
- */
-data class Card(val value: Int, val suit: Suits, val faceUp: Boolean = false) {
-
-    override fun toString(): String = if (faceUp) "${cardsMap[value]} of ${suit.suit}" else "???"
-}
 
 /**
  *  Composable that displays a [Card]. Depending on given [card]'s faceUp value will either display
@@ -84,7 +76,7 @@ fun SolitaireCard(
                         modifier = Modifier
                             .weight(1f)
                             .fillMaxSize()
-                        )
+                    )
                 }
                 Image(
                     painter = painterResource(card.suit.icon),
@@ -111,27 +103,41 @@ fun SolitaireCardFaceUpPreview() {
         Row(
             modifier = Modifier.fillMaxSize()
         ) {
-            SolitaireCard(Card(0, Suits.CLUBS, faceUp = true), modifier = Modifier
-                .weight(1f)
-                .fillMaxHeight(0.196f))
-            SolitaireCard(Card(12, Suits.DIAMONDS, faceUp = true), modifier = Modifier
-                .weight(1f)
-                .fillMaxHeight(0.196f))
-            SolitaireCard(Card(11, Suits.HEARTS, faceUp = true), modifier = Modifier
-                .weight(1f)
-                .fillMaxHeight(0.196f))
-            SolitaireCard(Card(10, Suits.SPADES, faceUp = true), modifier = Modifier
-                .weight(1f)
-                .fillMaxHeight(0.196f))
-            SolitaireCard(Card(9, Suits.CLUBS, faceUp = true), modifier = Modifier
-                .weight(1f)
-                .fillMaxHeight(0.196f))
-            SolitaireCard(Card(8, Suits.DIAMONDS, faceUp = true), modifier = Modifier
-                .weight(1f)
-                .fillMaxHeight(0.196f))
-            SolitaireCard(Card(7, Suits.HEARTS, faceUp = true), modifier = Modifier
-                .weight(1f)
-                .fillMaxHeight(0.196f))
+            SolitaireCard(
+                Card(0, Suits.CLUBS, faceUp = true), modifier = Modifier
+                    .weight(1f)
+                    .fillMaxHeight(0.196f)
+            )
+            SolitaireCard(
+                Card(12, Suits.DIAMONDS, faceUp = true), modifier = Modifier
+                    .weight(1f)
+                    .fillMaxHeight(0.196f)
+            )
+            SolitaireCard(
+                Card(11, Suits.HEARTS, faceUp = true), modifier = Modifier
+                    .weight(1f)
+                    .fillMaxHeight(0.196f)
+            )
+            SolitaireCard(
+                Card(10, Suits.SPADES, faceUp = true), modifier = Modifier
+                    .weight(1f)
+                    .fillMaxHeight(0.196f)
+            )
+            SolitaireCard(
+                Card(9, Suits.CLUBS, faceUp = true), modifier = Modifier
+                    .weight(1f)
+                    .fillMaxHeight(0.196f)
+            )
+            SolitaireCard(
+                Card(8, Suits.DIAMONDS, faceUp = true), modifier = Modifier
+                    .weight(1f)
+                    .fillMaxHeight(0.196f)
+            )
+            SolitaireCard(
+                Card(7, Suits.HEARTS, faceUp = true), modifier = Modifier
+                    .weight(1f)
+                    .fillMaxHeight(0.196f)
+            )
         }
     }
 }
@@ -143,27 +149,41 @@ fun SolitaireCardFaceDownPreview() {
         Row(
             modifier = Modifier.fillMaxSize()
         ) {
-            SolitaireCard(Card(0, Suits.CLUBS), modifier = Modifier
-                .weight(1f)
-                .fillMaxHeight(0.196f))
-            SolitaireCard(Card(12, Suits.DIAMONDS), modifier = Modifier
-                .weight(1f)
-                .fillMaxHeight(0.196f))
-            SolitaireCard(Card(11, Suits.HEARTS), modifier = Modifier
-                .weight(1f)
-                .fillMaxHeight(0.196f))
-            SolitaireCard(Card(10, Suits.SPADES), modifier = Modifier
-                .weight(1f)
-                .fillMaxHeight(0.196f))
-            SolitaireCard(Card(9, Suits.CLUBS), modifier = Modifier
-                .weight(1f)
-                .fillMaxHeight(0.196f))
-            SolitaireCard(Card(8, Suits.DIAMONDS), modifier = Modifier
-                .weight(1f)
-                .fillMaxHeight(0.196f))
-            SolitaireCard(Card(7, Suits.HEARTS), modifier = Modifier
-                .weight(1f)
-                .fillMaxHeight(0.196f))
+            SolitaireCard(
+                Card(0, Suits.CLUBS), modifier = Modifier
+                    .weight(1f)
+                    .fillMaxHeight(0.196f)
+            )
+            SolitaireCard(
+                Card(12, Suits.DIAMONDS), modifier = Modifier
+                    .weight(1f)
+                    .fillMaxHeight(0.196f)
+            )
+            SolitaireCard(
+                Card(11, Suits.HEARTS), modifier = Modifier
+                    .weight(1f)
+                    .fillMaxHeight(0.196f)
+            )
+            SolitaireCard(
+                Card(10, Suits.SPADES), modifier = Modifier
+                    .weight(1f)
+                    .fillMaxHeight(0.196f)
+            )
+            SolitaireCard(
+                Card(9, Suits.CLUBS), modifier = Modifier
+                    .weight(1f)
+                    .fillMaxHeight(0.196f)
+            )
+            SolitaireCard(
+                Card(8, Suits.DIAMONDS), modifier = Modifier
+                    .weight(1f)
+                    .fillMaxHeight(0.196f)
+            )
+            SolitaireCard(
+                Card(7, Suits.HEARTS), modifier = Modifier
+                    .weight(1f)
+                    .fillMaxHeight(0.196f)
+            )
         }
     }
 }
