+++ app/src/main/java/com/heyzeusv/solitaire/ui/Pile.kt
@@ -1,4 +1,4 @@
-package com.heyzeusv.solitaire
+package com.heyzeusv.solitaire.ui
 
 import androidx.annotation.DrawableRes
 import androidx.compose.foundation.Image
@@ -8,47 +8,10 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
+import com.heyzeusv.solitaire.R
+import com.heyzeusv.solitaire.data.Card
 import com.heyzeusv.solitaire.util.SolitairePreview
-
-/**
- *  Class that handles 52 [Card] deck creation.
- */
-class Stock(initialPile: List<Card> = emptyList()) : Pile(initialPile) {
-
-    /**
-     *  Add given [cards] to [mPile].
-     */
-    override fun add(cards: List<Card>): Boolean {
-        return mPile.addAll(cards.map { it.copy(faceUp = false) })
-    }
-
-    /**
-     *  Remove the first [Card] in [mPile] and return it.
-     */
-    override fun remove(tappedIndex: Int): Card = mPile.removeFirst()
-
-    /**
-     *  Reset [mPile] using given [cards].
-     */
-    override fun reset(cards: List<Card>) {
-        mPile.clear()
-        add(cards)
-    }
-
-    /**
-     *  Used to return [mPile] to a previous state of given [cards].
-     */
-    override fun undo(cards: List<Card>) {
-        mPile.clear()
-        if (cards.isEmpty()) return
-        // only last card is shown to user, this makes sure it is not visible
-        val mutableCards = cards.toMutableList()
-        mutableCards[mutableCards.size - 1] = mutableCards.last().copy(faceUp = false)
-        mPile.addAll(mutableCards)
-    }
-
-    override fun toString(): String = pile.toList().toString()
-}
+import com.heyzeusv.solitaire.util.Suits
 
 /**
  *  Composable that displays [pile]. If given [pile] is empty, [emptyIconId] is displayed.
